panel/panel_coverage/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
crcmod==1.7
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
idna==3.3
mypy==0.931
mypy-extensions==0.4.3
numpy==1.19.5
pandas==1.1.5
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
tomli==2.0.0
typed-ast==1.5.2
typing-extensions==4.0.1
urllib3==1.26.7